requirements.txt
aioredis==2.0.1
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
backoff==2.2.1
bcrypt==4.2.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.51
botocore==1.35.51
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.4
fastapi-mail==1.4.1
filelock==3.16.1
filetype==1.2.0
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
langdetect==1.0.9
layoutparser==0.0.1
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy<2.0
langchain==0.3.6
# numpy==2.1.2
onnx==1.17.0
opencv-python==4.6.0.66
packaging==24.1
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20231228
pillow==11.0.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pycparser==2.22
tesseract==0.1.3
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pypdf>=4.0.1,<5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.16
PyYAML==6.0.2
RapidFuzz==3.10.1
redis==5.2.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.1
starlette==0.41.2
tabulate==0.9.0
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.4
unstructured-client==0.26.2
unstructured_inference==0.2.4
urllib3==1.25.4
uvicorn==0.32.0
wrapt==1.16.0
python-pptx==1.0.2
XlsxWriter==3.2.0
langchain-community==0.3.5
pytesseract==0.3.13
openai==1.53.0
itsdangerous==2.2.0